refactor: less out/src
