+++ packages/salesforcedx-visualforce-language-server/src/modes/embeddedSupport.ts
@@ -5,7 +5,7 @@
 
 'use strict';
 
-import { LanguageService, TokenType } from '@salesforce/salesforcedx-visualforce-markup-language-server/out/src';
+import { LanguageService, TokenType } from '@salesforce/salesforcedx-visualforce-markup-language-server';
 import { Position, Range, TextDocument } from 'vscode-languageserver-types';
 
 type LanguageRange = Range & {
