+++ packages/salesforcedx-visualforce-language-server/src/visualforceServer.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { DocumentContext } from '@salesforce/salesforcedx-visualforce-markup-language-server/out/src';
+import { DocumentContext } from '@salesforce/salesforcedx-visualforce-markup-language-server';
 import * as path from 'node:path';
 import * as url from 'node:url';
 import {
