+++ packages/salesforcedx-visualforce-language-server/test/jest/embedded.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { getLanguageService } from '@salesforce/salesforcedx-visualforce-markup-language-server/out/src';
+import { getLanguageService } from '@salesforce/salesforcedx-visualforce-markup-language-server';
 import * as assert from 'node:assert';
 import { TextDocument } from 'vscode-languageserver-types';
 import * as embeddedSupport from '../../src/modes/embeddedSupport';
